Switch IdentityServer for E2E tests
IdentityServer4 has become Duende Identity Server, and they switched
over the demo instance to a new location. They also tweaked some
html on the login form requiring changes to the tests.

It's tempting to switch over to a locally running NPM based OIDC server
for stability, but then again the Duende folks are so great at building
and running IDS that it seems worth it to stick with that one as long
as it conveniently lasts.

Fixes #109